admin_dash/B2B-Freshk-backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
black==24.2.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
coverage==7.4.1
dj-database-url
Django==5.1.3
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.3.0
frozenlist==1.6.0
gunicorn==21.2.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
multidict==6.4.4
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
propcache==0.3.1
psycopg2-binary
PyJWT==2.10.1
pytest==8.0.0
pytest-django==4.8.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
setuptools
sqlparse==0.5.3
tomli==2.2.1
twilio==9.6.2
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
whitenoise==6.6.0
yarl==1.20.0
